reorganize streams to individual files
